Add onResized and onDestroyed signals
Added a onResized signal that is sent every time the size of the element
changes. Container now uses this signal of its child elements to connect
its calculateSize function to.
Added onDestroyed signal that is sent when the destructor of the element
is called. Used by Container to remove child elements that are destroyed